% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stph_palette}
\alias{stph_palette}
\title{Define the Swiss TPH color palette}
\usage{
stph_palette()
}
\value{
A vector containing color values for different shades.
}
\description{
This function defines a color palette with carefully selected Swiss Tropical
and Public Health Institute (Swiss TPH) color values. The palette includes
various shades of red, blue, green, yellow, and grey, providing a visually
harmonious set of colors for data visualization and presentation purposes.
}
