% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{vertical_boxplot}
\alias{vertical_boxplot}
\title{Generate a vertical boxplot for a given data frame, continuous variable, and categorical variable.}
\usage{
vertical_boxplot(df, continuous_var, categorical_var)
}
\arguments{
\item{df}{A data frame containing the variables of interest.}

\item{continuous_var}{A character string specifying the name of the continuous variable.}

\item{categorical_var}{A character string specifying the name of the categorical variable.}
}
\value{
A ggplot2 object representing the vertical boxplot.
}
\description{
This function creates a vertical boxplot using ggplot2 for visualizing the distribution
of a continuous variable across different categories.
}
