% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_ppt_report}
\alias{generate_ppt_report}
\title{Generate a Powerpoint presentation using Quarto}
\usage{
generate_ppt_report(
  qmd_filename,
  pptx_filename,
  filename_format = "",
  params = list()
)
}
\arguments{
\item{qmd_filename}{Filename of the Quarto document to be processed (with *.qmd extension).}

\item{pptx_filename}{Filename for the generated Quarto-rendered presentation (without *.pptx extension).}

\item{filename_format}{Specify the filename format, options include an empty string (default) or "with_date".}

\item{params}{List of parameters to customize the report generation process (optional, default is an empty list).}
}
\value{
The function does not explicitly return a value but generates a Powerpoint presentation in the working directory.
}
\description{
This function generates a standardized Quarto document in PPTX format.
}
