% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_word_report}
\alias{generate_word_report}
\title{Generate a Microsoft Word report using Quarto}
\usage{
generate_word_report(
  report_dir,
  quarto_filename,
  report_filename,
  params = list()
)
}
\arguments{
\item{report_dir}{Path to the folder where the generated DOCX Quarto report is stored}

\item{quarto_filename}{Filename of the Quarto document to run}

\item{report_filename}{Filename of the Quarto rendered report}

\item{params}{List of parameters}
}
\description{
This function generates a standardised Quarto document in a DOCX format
}
