% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_word_report}
\alias{generate_word_report}
\title{Generate a Microsoft Word report using Quarto}
\usage{
generate_word_report(
  qmd_filename,
  docx_filename,
  filename_format = "",
  params = list()
)
}
\arguments{
\item{qmd_filename}{Filename of the Quarto document to be processed (with *.qmd extension).}

\item{docx_filename}{Filename for the generated Quarto-rendered report (without *.docx extension).}

\item{filename_format}{Specify the filename format, options include an empty string (default) or "with_date".}

\item{params}{List of parameters to customize the report generation process (optional, default is an empty list).}
}
\value{
The function does not explicitly return a value but generates a Quarto report in the working directory.
}
\description{
This function generates a standardized Quarto document in DOCX format.
}
